+++ Classes/Controller/AbstractCommandController.php
@@ -1,7 +1,7 @@
 <?php
 namespace T3v\T3vCore\Controller;
 
-use TYPO3\CMS\Extbase\Mvc\Controller\CommandController as CommandControllerExtbase;
+use TYPO3\CMS\Extbase\Mvc\Controller\CommandController;
 
 use Colors\Color;
 
@@ -10,7 +10,7 @@ use Colors\Color;
  *
  * @package T3v\T3vCore\Controller
  */
-abstract class AbstractCommandController extends CommandControllerExtbase {
+abstract class AbstractCommandController extends CommandController {
   /**
    * Logs respectively echos a message.
    *